change configuration connection, entity configurations
