+++ CityTransportSystem/DataAccess/Entities/Stop.cs
