+++ CityTransportSystem/DataAccess/Entities/StopRouteTrip.cs
