+++ CityTransportSystem/DataAccess/Entities/Ticket.cs
