+++ CityTransportSystem/DataAccess/Entities/TransportTypeStop.cs
