+++ CityTransportSystem/DataAccess/Entities/Trip.cs
