+++ CityTransportSystem/DataAccess/Entities/TripTransport.cs
