+++ CityTransportSystem/DataAccess/Entities/TripValidation.cs
