+++ CityTransportSystem/DataAccess/EntityConfigurations/StopConfiguration.cs
