+++ CityTransportSystem/DataAccess/EntityConfigurations/StopRouteTripConfiguration.cs
