+++ CityTransportSystem/DataAccess/EntityConfigurations/TicketConfiguration.cs
