+++ CityTransportSystem/DataAccess/EntityConfigurations/TransportTypeConfiguration.cs
