+++ CityTransportSystem/DataAccess/EntityConfigurations/TransportTypeStopConfiguration.cs
