+++ CityTransportSystem/DataAccess/EntityConfigurations/TripConfiguration.cs
