+++ CityTransportSystem/DataAccess/EntityConfigurations/TripTransportConfiguration.cs
