+++ CityTransportSystem/DataAccess/EntityConfigurations/TripValidationConfiguration.cs
