+++ CityTransportSystem/WebAPI/Startup.cs
@@ -4,9 +4,11 @@ using CTSWebAPI.Mappings;
 using FluentValidation.AspNetCore;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using System.IO;
 
 namespace CTSWebAPI
 {
@@ -22,8 +24,12 @@ namespace CTSWebAPI
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddDbContext<AppDbContext>()
-                    .AddAutoMapper(map => map.AddProfile<MapingProfile>(), typeof(Startup))
+            services.AddDbContext<AppDbContext>(options => 
+                    {                   
+                        string connectionString = Configuration.GetConnectionString("PostgreSQL");
+                        options.UseNpgsql(connectionString);
+                    })
+                    .AddAutoMapper(map => map.AddProfile<MappingProfile>(), typeof(Startup))
                     .AddControllers()
                     .AddFluentValidation()
                     .AddNewtonsoftJson(options =>
