+++ CityTransportSystem/WebAPI/WebAPI.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CTSCore\CTSCore.csproj" />
+    <ProjectReference Include="..\CTSCore\DataAccess.csproj" />
   </ItemGroup>
